change type of point to int (reverted from commit 2eb1a935baab919acb784321cd6d8a3c04676404)

+++ src/java/linkjvm/high/vision/Point2.java
@@ -1,15 +1,15 @@
 package linkjvm.high.vision;
 
 public class Point2 {
-	private int x;
-	private int y;
+	private long x;
+	private long y;
 	
 	/**
 	 * 
 	 * @param x
 	 * @param y
 	 */
-	public Point2(int x, int y) {
+	public Point2(long x, long y) {
 		super();
 		this.x = x;
 		this.y = y;
@@ -18,28 +18,28 @@ public class Point2 {
 	/**
 	 * @return the x
 	 */
-	public synchronized int getX() {
+	public synchronized long getX() {
 		return x;
 	}
 
 	/**
 	 * @param x the x to set
 	 */
-	public synchronized void setX(int x) {
+	public synchronized void setX(long x) {
 		this.x = x;
 	}
 
 	/**
 	 * @return the y
 	 */
-	public synchronized int getY() {
+	public synchronized long getY() {
 		return y;
 	}
 
 	/**
 	 * @param y the y to set
 	 */
-	public synchronized void setY(int y) {
+	public synchronized void setY(long y) {
 		this.y = y;
 	}
